+++ code/src/ui/graphic/table/pagination/table-pagination.tsx
@@ -1,9 +1,8 @@
 import { Table } from "@tanstack/react-table";
-
-import { FieldValues } from "react-hook-form/dist/types/fields";
-import Button from "ui/interaction/button";
 import { useDictionary } from "hooks/useDictionary";
+import { FieldValues } from "react-hook-form/dist/types/fields";
 import { Dictionary } from "types/dictionary";
+import Button from "ui/interaction/button";
 
 interface TablePaginationProps {
   table: Table<FieldValues>;
@@ -20,11 +19,12 @@ const PAGE_SIZE_OPTIONS: number[] = [5, 10, 20, 50];
 export default function TablePagination(props: Readonly<TablePaginationProps>) {
   const dict: Dictionary = useDictionary();
   const { table } = props;
+  const numberOfSelectedRows: number = table.getSelectedRowModel().rows.length;
   return (
     <div className="flex items-center justify-between p-4 bg-muted border-t border-border flex-shrink-0">
       <div className="text-sm text-foreground">
         {dict.message.numberOfRecords
-          .replace("{replace}", String(table.getFilteredRowModel().rows.length))
+          .replace("{replace}", String(numberOfSelectedRows > 0 ? numberOfSelectedRows : table.getFilteredRowModel().rows.length))
           .replace(
             "{replacetotal}",
             String(table.getCoreRowModel().rows.length)